pyproject.toml
[tool.poetry]
name = "django-dramatiq-pg"
version = "1.3.2"
description = "Integration of Django with dramatiq-pg"
readme = "README.rst"
authors = ["Curtis Maloney <curtis@tinbrain.net>"]
license = "MIT"
homepage = "https://github.com/uptick/django-dramatiq-pg"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Framework :: Django",
  "Framework :: Django :: 3.0",
  "Framework :: Django :: 3.1",
]

[tool.poetry.dependencies]
python = "^3.7"
django = ">= 3.1"
dramatiq-pg = ">= 0.9.0"

[tool.poetry.dev-dependencies]
psycopg2 = "^2.8.4"
flake8 = "^3.7.9"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
isort = "^5.11.4"

[tool.isort]
known_django = "django"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]
extra_standard_library = ["arrow", "dateutil", "factory", "invoke", "pytz", "requests", "tablib", "yaml"]
multi_line_output = 3
line_length = 100
include_trailing_comma = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'app.settings'
pythonpath = "src tests"
testpaths = "tests"
